+++ src/Recyclarr.TrashLib.Tests/CustomFormat/Processors/PersistenceSteps/CustomFormatApiPersistenceStepTest.cs
@@ -14,7 +14,7 @@ public class CustomFormatApiPersistenceStepTest
 {
     private static ProcessedCustomFormatData QuickMakeCf(string cfName, string trashId, int cfId)
     {
-        return NewCf.Processed(cfName, trashId, new TrashIdMapping(trashId, cfName) {CustomFormatId = cfId});
+        return NewCf.Processed(cfName, trashId, cfId);
     }
 
     [Test]
@@ -24,7 +24,7 @@ public class CustomFormatApiPersistenceStepTest
         transactions.NewCustomFormats.Add(QuickMakeCf("cfname1", "trashid1", 1));
         transactions.UpdatedCustomFormats.Add(QuickMakeCf("cfname2", "trashid2", 2));
         transactions.UnchangedCustomFormats.Add(QuickMakeCf("cfname3", "trashid3", 3));
-        transactions.DeletedCustomFormatIds.Add(new TrashIdMapping("trashid4", "cfname4") {CustomFormatId = 4});
+        transactions.DeletedCustomFormatIds.Add(new TrashIdMapping("trashid4", "cfname4", 4));
 
         var api = Substitute.For<ICustomFormatService>();