+++ src/Recyclarr.TrashLib/Services/CustomFormat/Models/Cache/CustomFormatCache.cs
@@ -4,7 +4,7 @@ using Recyclarr.TrashLib.Cache;
 namespace Recyclarr.TrashLib.Services.CustomFormat.Models.Cache;
 
 [CacheObjectName("custom-format-cache")]
-public class CustomFormatCache
+public record CustomFormatCache
 {
     public const int LatestVersion = 1;
 
@@ -12,16 +12,4 @@ public class CustomFormatCache
     public Collection<TrashIdMapping> TrashIdMappings { get; init; } = new();
 }
 
-public class TrashIdMapping
-{
-    public TrashIdMapping(string trashId, string customFormatName, int customFormatId = default)
-    {
-        TrashId = trashId;
-        CustomFormatName = customFormatName;
-        CustomFormatId = customFormatId;
-    }
-
-    public string TrashId { get; }
-    public string CustomFormatName { get; }
-    public int CustomFormatId { get; set; }
-}
+public record TrashIdMapping(string TrashId, string CustomFormatName, int CustomFormatId);