+++ src/Recyclarr.TrashLib/Services/CustomFormat/Models/ProcessedCustomFormatData.cs
@@ -1,6 +1,4 @@
-using System.Diagnostics.CodeAnalysis;
 using Newtonsoft.Json.Linq;
-using Recyclarr.TrashLib.Services.CustomFormat.Models.Cache;
 
 namespace Recyclarr.TrashLib.Services.CustomFormat.Models;
 
@@ -18,17 +16,5 @@ public class ProcessedCustomFormatData
     public string TrashId => _data.TrashId;
     public int? Score => _data.Score;
     public JObject Json { get; set; }
-    public TrashIdMapping? CacheEntry { get; set; }
-
-    public void SetCache(int customFormatId)
-    {
-        // Do not pass the customFormatId to constructor since an instance may already exist.
-        CacheEntry ??= new TrashIdMapping(TrashId, Name);
-        CacheEntry.CustomFormatId = customFormatId;
-    }
-
-    [SuppressMessage("Microsoft.Design", "CA1024", Justification = "Method throws an exception")]
-    public int GetCustomFormatId()
-        => CacheEntry?.CustomFormatId ??
-            throw new InvalidOperationException("CacheEntry must exist to obtain custom format ID");
+    public int FormatId { get; set; }
 }